+++ invenio_rdm_records/assets/semantic-ui/js/invenio_rdm_records/src/deposit/fields/CopyrightsField/CopyrightsField.js
@@ -8,39 +8,72 @@
 
 import React, { Component } from "react";
 import PropTypes from "prop-types";
-
-import { FieldLabel, TextField } from "react-invenio-forms";
+import Overridable from "react-overridable";
 import { i18next } from "@translations/invenio_rdm_records/i18next";
+import {
+  showHideOverridable,
+  fieldCommonProps,
+  FieldLabel,
+  TextField,
+} from "react-invenio-forms";
 
-export class CopyrightsField extends Component {
+class CopyrightsFieldComponent extends Component {
   render() {
-    const { fieldPath, label, required } = this.props;
+    const {
+      fieldPath,
+      label,
+      labelIcon,
+      required,
+      disabled,
+      helpText,
+      placeholder,
+      optimized,
+    } = this.props;
+
     return (
-      <TextField
-        fieldPath={fieldPath}
-        label={
-          <FieldLabel htmlFor={fieldPath} icon="copyright outline" label={label} />
-        }
+      <Overridable
+        id="InvenioRdmRecords.DepositForm.CopyrightsField.Container"
+        label={label}
+        labelIcon={labelIcon}
         required={required}
-        helpText={i18next.t(
-          "A copyright statement describing the ownership of the uploaded resource."
-        )}
-        placeholder={i18next.t("Copyright (C) {{currentYear}} The Authors.", {
-          currentYear: new Date().getFullYear(),
-        })}
-        optimized
-      />
+        disabled={disabled}
+        helpText={helpText}
+        placeholder={placeholder}
+        optimized={optimized}
+      >
+        <TextField
+          fieldPath={fieldPath}
+          label={<FieldLabel htmlFor={fieldPath} icon={labelIcon} label={label} />}
+          required={required}
+          disabled={disabled}
+          helpText={helpText}
+          placeholder={placeholder}
+          optimized={optimized}
+        />
+      </Overridable>
     );
   }
 }
 
-CopyrightsField.propTypes = {
-  fieldPath: PropTypes.string.isRequired,
-  label: PropTypes.string,
-  required: PropTypes.bool,
+CopyrightsFieldComponent.propTypes = {
+  optimized: PropTypes.bool,
+  ...fieldCommonProps,
 };
 
-CopyrightsField.defaultProps = {
+CopyrightsFieldComponent.defaultProps = {
   label: i18next.t("Copyright"),
+  labelIcon: "copyright outline",
   required: false,
+  helpText: i18next.t(
+    "A copyright statement describing the ownership of the uploaded resource."
+  ),
+  placeholder: i18next.t("Copyright (C) {{currentYear}} The Authors.", {
+    currentYear: new Date().getFullYear(),
+  }),
+  optimized: true,
 };
+
+export const CopyrightsField = showHideOverridable(
+  "InvenioRdmRecords.DepositForm.CopyrightsField",
+  CopyrightsFieldComponent
+);