+++ invenio_rdm_records/assets/semantic-ui/js/invenio_rdm_records/src/deposit/fields/DatesField/DatesField.js
@@ -8,8 +8,15 @@
 
 import React, { Component } from "react";
 import PropTypes from "prop-types";
-import { ArrayField, GroupField, SelectField, TextField } from "react-invenio-forms";
-import { Button, Form, Icon } from "semantic-ui-react";
+import {
+  ArrayField,
+  GroupField,
+  SelectField,
+  TextField,
+  showHideOverridable,
+  fieldCommonProps,
+} from "react-invenio-forms";
+import { Button, Form } from "semantic-ui-react";
 import _isEmpty from "lodash/isEmpty";
 import _matches from "lodash/matches";
 import _filter from "lodash/filter";
@@ -20,7 +27,7 @@ import { i18next } from "@translations/invenio_rdm_records/i18next";
 import { sortOptions } from "../../utils";
 import Overridable from "react-overridable";
 
-export class DatesField extends Component {
+class DatesFieldComponent extends Component {
   /** Top-level Dates Component */
 
   /**
@@ -47,24 +54,35 @@ export class DatesField extends Component {
       options,
       label,
       labelIcon,
-      placeholderDate,
       required,
       requiredOptions,
       showEmptyValue,
+      placeholder,
+      helpText,
+      addButtonLabel,
+      optimized,
     } = this.props;
 
     return (
       <Overridable
-        id="InvenioRdmRecords.DatesField.AddDateArrayField.Container"
+        id="InvenioRdmRecords.DepositForm.DatesField.Container"
         fieldPath={fieldPath}
+        defaultNewValue={emptyDate}
+        label={label}
+        labelIcon={labelIcon}
+        required={required}
+        requiredOptions={requiredOptions}
+        showEmptyValue={showEmptyValue}
+        placeholder={placeholder}
+        helpText={helpText}
+        addButtonLabel={addButtonLabel}
+        optimized={optimized}
       >
         <ArrayField
-          addButtonLabel={i18next.t("Add date")} // TODO: Pass by prop
+          addButtonLabel={addButtonLabel}
           defaultNewValue={emptyDate}
           fieldPath={fieldPath}
-          helpText={i18next.t(
-            "Format: DATE or DATE/DATE where DATE is YYYY or YYYY-MM or YYYY-MM-DD."
-          )}
+          helpText={helpText}
           label={label}
           labelIcon={labelIcon}
           required={required}
@@ -78,22 +96,22 @@ export class DatesField extends Component {
             const hasRequiredTypeValue = _has(requiredOption, "type");
             const hasRequiredDescriptionValue = _has(requiredOption, "description");
             return (
-              <GroupField fieldPath={fieldPath} optimized>
+              <GroupField fieldPath={fieldPath} optimized={optimized}>
                 <Overridable
-                  id="InvenioRdmRecords.DatesField.DateTextField.Container"
+                  id="InvenioRdmRecords.DepositForm.DatesField.DateField"
                   fieldPath={`${fieldPathPrefix}.date`}
                 >
                   <TextField
                     fieldPath={`${fieldPathPrefix}.date`}
                     label={i18next.t("Date")}
-                    placeholder={placeholderDate}
+                    placeholder={placeholder}
                     disabled={hasRequiredDateValue}
                     required
                     width={5}
                   />
                 </Overridable>
                 <Overridable
-                  id="InvenioRdmRecords.DatesField.TypeSelectField.Container"
+                  id="InvenioRdmRecords.DepositForm.DatesField.TypeField"
                   fieldPath={`${fieldPathPrefix}.type`}
                 >
                   <SelectField
@@ -104,12 +122,13 @@ export class DatesField extends Component {
                     disabled={hasRequiredTypeValue}
                     required
                     width={5}
-                    optimized
+                    optimized={optimized}
                   />
                 </Overridable>
                 <Overridable
-                  id="InvenioRdmRecords.DatesField.DescriptionTextField.Container"
-                  fieldPath={`${fieldPathPrefix}.description`}
+                  id="InvenioRdmRecords.DepositForm.DatesField.DescriptionField"
+                  disabled={hasRequiredDescriptionValue}
+                  width={5}
                 >
                   <TextField
                     fieldPath={`${fieldPathPrefix}.description`}
@@ -118,18 +137,16 @@ export class DatesField extends Component {
                     width={5}
                   />
                 </Overridable>
-                <Overridable id="InvenioRdmRecords.DatesField.RemoveFormField.Container">
+                <Overridable id="InvenioRdmRecords.DepositForm.DatesField.RemoveButton">
                   <Form.Field>
                     <Button
                       aria-label={i18next.t("Remove field")}
                       className="close-btn"
                       disabled={!_isEmpty(requiredOption)}
-                      icon
+                      icon="close"
                       onClick={() => arrayHelpers.remove(indexPath)}
                       type="button"
-                    >
-                      <Icon name="close" />
-                    </Button>
+                    />
                   </Form.Field>
                 </Overridable>
               </GroupField>
@@ -141,10 +158,7 @@ export class DatesField extends Component {
   }
 }
 
-DatesField.propTypes = {
-  fieldPath: PropTypes.string.isRequired,
-  label: PropTypes.string,
-  labelIcon: PropTypes.string,
+DatesFieldComponent.propTypes = {
   options: PropTypes.shape({
     type: PropTypes.arrayOf(
       PropTypes.shape({
@@ -153,17 +167,27 @@ DatesField.propTypes = {
       })
     ),
   }).isRequired,
-  required: PropTypes.bool,
-  placeholderDate: PropTypes.string,
   requiredOptions: PropTypes.array,
   showEmptyValue: PropTypes.bool,
+  optimized: PropTypes.bool,
+  ...fieldCommonProps,
 };
 
-DatesField.defaultProps = {
+DatesFieldComponent.defaultProps = {
   label: i18next.t("Dates"),
   labelIcon: "calendar",
-  placeholderDate: i18next.t("YYYY-MM-DD or YYYY-MM-DD/YYYY-MM-DD"),
+  placeholder: i18next.t("YYYY-MM-DD or YYYY-MM-DD/YYYY-MM-DD"),
+  helpText: i18next.t(
+    "Format: DATE or DATE/DATE where DATE is YYYY or YYYY-MM or YYYY-MM-DD."
+  ),
+  addButtonLabel: i18next.t("Add date"),
   required: false,
   requiredOptions: [],
   showEmptyValue: false,
+  optimized: true,
 };
+
+export const DatesField = showHideOverridable(
+  "InvenioRdmRecords.DepositForm.DatesField",
+  DatesFieldComponent
+);