+++ invenio_rdm_records/assets/semantic-ui/js/invenio_rdm_records/src/deposit/fields/PublisherField/PublisherField.js
@@ -8,36 +8,60 @@
 
 import React, { Component } from "react";
 import PropTypes from "prop-types";
-
-import { FieldLabel, TextField } from "react-invenio-forms";
+import Overridable from "react-overridable";
+import {
+  FieldLabel,
+  TextField,
+  showHideOverridable,
+  fieldCommonProps,
+} from "react-invenio-forms";
 import { i18next } from "@translations/invenio_rdm_records/i18next";
 
-export class PublisherField extends Component {
+class PublisherFieldComponent extends Component {
   render() {
-    const { fieldPath, label, labelIcon, placeholder } = this.props;
+    const { fieldPath, label, labelIcon, placeholder, disabled, required, helpText } =
+      this.props;
 
     return (
-      <TextField
+      <Overridable
+        id="InvenioRdmRecords.DepositForm.PublisherField.Container"
         fieldPath={fieldPath}
-        helpText={i18next.t(
-          "The publisher is used to formulate the citation, so consider the prominence of the role."
-        )}
-        label={<FieldLabel htmlFor={fieldPath} icon={labelIcon} label={label} />}
+        icon={labelIcon}
+        label={label}
         placeholder={placeholder}
-      />
+        disabled={disabled}
+        required={required}
+        helpText={helpText}
+      >
+        <TextField
+          fieldPath={fieldPath}
+          helpText={helpText}
+          label={<FieldLabel htmlFor={fieldPath} icon={labelIcon} label={label} />}
+          placeholder={placeholder}
+          disabled={disabled}
+          required={required}
+        />
+      </Overridable>
     );
   }
 }
 
-PublisherField.propTypes = {
+PublisherFieldComponent.propTypes = {
   fieldPath: PropTypes.string.isRequired,
-  label: PropTypes.string,
-  labelIcon: PropTypes.string,
   placeholder: PropTypes.string,
+  ...fieldCommonProps,
 };
 
-PublisherField.defaultProps = {
+PublisherFieldComponent.defaultProps = {
   label: i18next.t("Publisher"),
   labelIcon: "building outline",
   placeholder: i18next.t("Publisher"),
+  helpText: i18next.t(
+    "The publisher is used to formulate the citation, so consider the prominence of the role."
+  ),
 };
+
+export const PublisherField = showHideOverridable(
+  "InvenioRdmRecords.DepositForm.PublisherField",
+  PublisherFieldComponent
+);