+++ invenio_rdm_records/assets/semantic-ui/js/invenio_rdm_records/src/deposit/fields/ReferencesField/ReferencesField.js
@@ -8,67 +8,101 @@
 
 import React, { Component } from "react";
 import PropTypes from "prop-types";
-
-import { TextField, GroupField, ArrayField, FieldLabel } from "react-invenio-forms";
-import { Button, Form, Icon } from "semantic-ui-react";
-
+import Overridable from "react-overridable";
+import {
+  TextField,
+  GroupField,
+  ArrayField,
+  FieldLabel,
+  showHideOverridable,
+  fieldCommonProps,
+} from "react-invenio-forms";
+import { Button, Form } from "semantic-ui-react";
 import { emptyReference } from "./initialValues";
 import { i18next } from "@translations/invenio_rdm_records/i18next";
 
-export class ReferencesField extends Component {
+class ReferencesFieldComponent extends Component {
   render() {
-    const { fieldPath, label, labelIcon, required, showEmptyValue } = this.props;
+    const {
+      fieldPath,
+      label,
+      labelIcon,
+      required,
+      disabled,
+      showEmptyValue,
+      helpText,
+      addButtonLabel,
+      optimized,
+    } = this.props;
 
     return (
-      <ArrayField
-        addButtonLabel={i18next.t("Add reference")}
+      <Overridable
+        id="InvenioRdmRecords.DepositForm.ReferencesField.Container"
         defaultNewValue={emptyReference}
-        fieldPath={fieldPath}
-        label={<FieldLabel htmlFor={fieldPath} icon={labelIcon} label={label} />}
         required={required}
+        disabled={disabled}
         showEmptyValue={showEmptyValue}
+        helpText={helpText}
+        addButtonLabel={addButtonLabel}
+        optimized={optimized}
       >
-        {({ arrayHelpers, indexPath }) => {
-          const fieldPathPrefix = `${fieldPath}.${indexPath}`;
+        <ArrayField
+          addButtonLabel={addButtonLabel}
+          defaultNewValue={emptyReference}
+          fieldPath={fieldPath}
+          label={<FieldLabel htmlFor={fieldPath} icon={labelIcon} label={label} />}
+          required={required}
+          disabled={disabled}
+          showEmptyValue={showEmptyValue}
+          helpText={helpText}
+        >
+          {({ arrayHelpers, indexPath }) => {
+            const fieldPathPrefix = `${fieldPath}.${indexPath}`;
 
-          return (
-            <GroupField optimized>
-              <TextField
-                fieldPath={`${fieldPathPrefix}.reference`}
-                label={i18next.t("Reference string")}
-                required
-                width={16}
-              />
+            return (
+              <GroupField optimized={optimized}>
+                <TextField
+                  fieldPath={`${fieldPathPrefix}.reference`}
+                  label={i18next.t("Reference string")}
+                  required
+                  width={16}
+                  optimized={optimized}
+                />
 
-              <Form.Field>
-                <Button
-                  aria-label={i18next.t("Remove field")}
-                  className="close-btn"
-                  icon
-                  onClick={() => arrayHelpers.remove(indexPath)}
-                >
-                  <Icon name="close" />
-                </Button>
-              </Form.Field>
-            </GroupField>
-          );
-        }}
-      </ArrayField>
+                <Form.Field>
+                  <Button
+                    aria-label={i18next.t("Remove field")}
+                    className="close-btn"
+                    icon="close"
+                    onClick={() => arrayHelpers.remove(indexPath)}
+                  />
+                </Form.Field>
+              </GroupField>
+            );
+          }}
+        </ArrayField>
+      </Overridable>
     );
   }
 }
 
-ReferencesField.propTypes = {
-  fieldPath: PropTypes.string.isRequired,
-  label: PropTypes.string,
-  labelIcon: PropTypes.string,
-  required: PropTypes.bool,
+ReferencesFieldComponent.propTypes = {
   showEmptyValue: PropTypes.bool,
+  addButtonLabel: PropTypes.string,
+  optimized: PropTypes.bool,
+  ...fieldCommonProps,
 };
 
-ReferencesField.defaultProps = {
+ReferencesFieldComponent.defaultProps = {
   label: i18next.t("References"),
   labelIcon: "bookmark",
   required: false,
   showEmptyValue: false,
+  addButtonLabel: i18next.t("Add reference"),
+  optimized: true,
 };
+
+export const ReferencesField = showHideOverridable(
+  "InvenioRdmRecords.DepositForm.ReferencesField",
+  ReferencesFieldComponent
+);