+++ invenio_rdm_records/assets/semantic-ui/js/invenio_rdm_records/src/deposit/fields/TitlesField/AdditionalTitlesField.js
@@ -9,28 +9,41 @@
 
 import React, { Component } from "react";
 import PropTypes from "prop-types";
-import { Button, Form, Icon } from "semantic-ui-react";
-
-import { ArrayField, GroupField, SelectField, TextField } from "react-invenio-forms";
+import { Button, Form } from "semantic-ui-react";
+import {
+  ArrayField,
+  GroupField,
+  SelectField,
+  TextField,
+  showHideOverridable,
+  fieldCommonProps,
+} from "react-invenio-forms";
 import { emptyAdditionalTitle } from "./initialValues";
 import { LanguagesField } from "../LanguagesField";
 import { i18next } from "@translations/invenio_rdm_records/i18next";
 
-export class AdditionalTitlesField extends Component {
+class AdditionalTitlesFieldComponent extends Component {
   render() {
-    const { fieldPath, options, recordUI } = this.props;
+    const { fieldPath, options, recordUI, helpText, addButtonLabel, optimized } =
+      this.props;
+
     return (
       <ArrayField
-        addButtonLabel={i18next.t("Add titles")}
+        addButtonLabel={addButtonLabel}
+        helpText={helpText}
         defaultNewValue={emptyAdditionalTitle}
         fieldPath={fieldPath}
         className="additional-titles"
       >
         {({ arrayHelpers, indexPath }) => {
           const fieldPathPrefix = `${fieldPath}.${indexPath}`;
+          const languagesInitialOptions =
+            recordUI?.additional_titles && recordUI.additional_titles[indexPath]?.lang
+              ? [recordUI.additional_titles[indexPath].lang]
+              : [];
 
           return (
-            <GroupField fieldPath={fieldPath} optimized>
+            <GroupField fieldPath={fieldPath} optimized={optimized}>
               <TextField
                 fieldPath={`${fieldPathPrefix}.title`}
                 label={i18next.t("Additional title")}
@@ -40,7 +53,7 @@ export class AdditionalTitlesField extends Component {
               <SelectField
                 fieldPath={`${fieldPathPrefix}.type`}
                 label={i18next.t("Type")}
-                optimized
+                optimized={optimized}
                 options={options.type}
                 required
                 width={5}
@@ -53,12 +66,7 @@ export class AdditionalTitlesField extends Component {
                     fieldPathPrefix: item.id,
                   }))
                 }
-                initialOptions={
-                  recordUI?.additional_titles &&
-                  recordUI.additional_titles[indexPath]?.lang
-                    ? [recordUI.additional_titles[indexPath].lang]
-                    : []
-                }
+                initialOptions={languagesInitialOptions}
                 fieldPath={`${fieldPathPrefix}.lang`}
                 label={i18next.t("Language")}
                 multiple={false}
@@ -68,15 +76,14 @@ export class AdditionalTitlesField extends Component {
                 selectOnBlur={false}
                 width={5}
               />
+
               <Form.Field>
                 <Button
                   aria-label={i18next.t("Remove field")}
                   className="close-btn"
-                  icon
+                  icon="close"
                   onClick={() => arrayHelpers.remove(indexPath)}
-                >
-                  <Icon name="close" />
-                </Button>
+                />
               </Form.Field>
             </GroupField>
           );
@@ -86,8 +93,7 @@ export class AdditionalTitlesField extends Component {
   }
 }
 
-AdditionalTitlesField.propTypes = {
-  fieldPath: PropTypes.string.isRequired,
+AdditionalTitlesFieldComponent.propTypes = {
   options: PropTypes.shape({
     type: PropTypes.arrayOf(
       PropTypes.shape({
@@ -104,9 +110,19 @@ AdditionalTitlesField.propTypes = {
     ),
   }),
   recordUI: PropTypes.object,
+  addButtonLabel: PropTypes.string,
+  optimized: PropTypes.bool,
+  ...fieldCommonProps,
 };
 
-AdditionalTitlesField.defaultProps = {
+AdditionalTitlesFieldComponent.defaultProps = {
   options: undefined,
   recordUI: undefined,
+  addButtonLabel: i18next.t("Add titles"),
+  optimized: true,
 };
+
+export const AdditionalTitlesField = showHideOverridable(
+  "InvenioRdmRecords.DepositForm.AdditionalTitlesField",
+  AdditionalTitlesFieldComponent
+);